+++ crates/action-runner/src/lib.rs
@@ -2,3 +2,4 @@ mod dep_graph;
 mod errors;
 
 pub use dep_graph::*;
+pub use errors::*;
